move alias tests to test_misc.py
